fix: resolve compilation errors and remove GPS text displays
Compilation Fixes:
- Replace FontAwesomeIcons.imageSlash with .image (icon exists)
- Fix distanceFilter type from double to int in LocationSettings

GPS Text Removal:
- Remove all 'Position capturée automatiquement' displays
- Remove GPS status indicators from checkout page
- Remove GPS status indicators from special order page
- GPS works silently in background without UI feedback

Payment Options:
- Remove 'Flutterwave' text from payment option
- Display only 'Carte bancaire' (cleaner UI)

Dialog Updates:
- Simplify driver assignment message
- Remove notification mention (cleaner message)

+++ lib/services/background_location_tracking_service.dart
@@ -20,7 +20,7 @@ class BackgroundLocationTrackingService {
   static const int _batchSize = 5;
   static const Duration _batchInterval = Duration(seconds: 30);
   static const Duration _updateInterval = Duration(seconds: 10);
-  static const double _distanceFilter = 10.0; // meters
+  static const int _distanceFilter = 10; // meters
   
   bool _isTracking = false;
   Position? _lastPosition;
